+++ system/src/sheets/FalloutNpcSheet.mjs
@@ -24,8 +24,19 @@ export default class FalloutNpcSheet extends FalloutBaseActorSheet {
 		return "systems/fallout/templates/actor/npc-sheet.hbs";
 	}
 
+	/** @override */
+	activateListeners(html) {
+		super.activateListeners(html);
+
+		html.find(".roll-wealth").click(this._onRollWealth.bind(this));
+	}
+
 	async getData(options) {
 		const context = await super.getData(options);
+
+		if (this.actor.isCreature) {
+			await this._prepareButcheryMaterials(context);
+		}
 
 		context.disableAutoXpReward = game.settings.get(
 			SYSTEM_ID, "disableAutoXpReward"
@@ -44,6 +55,24 @@ export default class FalloutNpcSheet extends FalloutBaseActorSheet {
 		}
 
 		return context;
+	}
+
+	async _onRollWealth(event) {
+		event.preventDefault();
+		const wealthLevel = Math.max(1, this.actor.system.wealth ?? 1);
+
+		const formula = `${wealthLevel}d20`;
+		const roll = new Roll(formula);
+
+		const wealthRoll = await roll.evaluate({ async: true });
+		try {
+			await game.dice3d.showForRoll(wealthRoll);
+		}
+		catch(err) {}
+
+		const caps = parseInt(roll.result);
+
+		this.actor.update({"system.currency.caps": caps});
 	}
 
 	async _updateObject(event, formData) {